Return errors on first entry of WalkDir
The first entry of WalkDir is being skipped - the problem with this is
that the first entry is where WalkDir can report that the directory for
comparison doesn't exist. The current code will return `false` for the
following cases:

  dir_diff::is_different("dir_that_exists", "dir_that_does_not_exist");
  dir_diff::is_different("dir_that_does_not_exist", "also_does_not_exist");

This is obviously not ideal since the directory not existing is probably
not what the user intended.

This commit fixes up the issue by checking the first entry returned from
WalkDir for error instead of skipping it. This causes `is_different` to
return an error if either or both directories don't actually exist.

Signed-off-by: Nick Stevens <nick@bitcurry.com>